@@ -8,32 +8,35 @@ from codecs import open
 from setuptools import setup, find_packages
 
 __title__ = 'tmtoolkit'
-__version__ = '0.10.0'
+__version__ = '0.11.0'
 __author__ = 'Markus Konrad'
 __license__ = 'Apache License 2.0'
 
 
 GITHUB_URL = 'https://github.com/WZBSocialScienceCenter/tmtoolkit'
 
-DEPS_BASE = ['numpy>=1.19.0,<2', 'scipy>=1.5.0,<1.6', 'pandas>=1.1.0,<1.2', 'xlrd>=1.2.0',
-             'globre>=0.1.5,<0.2', 'matplotlib>=3.3.0,<3.4', 'spacy>=2.3.0,<2.4']
+DEPS_BASE = ['numpy>=1.22.0', 'scipy>=1.7.0', 'globre>=0.1.5',
+             'pandas>=1.4.0', 'xlrd>=2.0.0', 'openpyxl>=3.0.0',
+             'matplotlib>=3.5.0']
 
 DEPS_EXTRA = {
-    'datatable': ['datatable>=0.10.0,<0.11'],
-    'nltk': ['nltk>=3.5.0,<3.6'],
-    'excel_export': ['openpyxl>=3.0.0'],
-    'wordclouds': ['wordcloud>=1.7.0,<1.8', 'Pillow>=7.2.0,<7.3'],
-    'lda': ['ldafork>=1.2.0,<1.3'],
-    'sklearn': ['scikit-learn>=0.23,<0.24'],
-    'gensim': ['gensim>=3.8.0,<3.9'],
-    'topic_modeling_eval_extra': ['gmpy2>=2.0.0,<3'],
-    'test': ['pytest>=6.0.0,<7', 'hypothesis>=5.23.0<5.24', 'decorator>=4.4.0,<4.5'],
-    'doc': ['Sphinx>=3.1.0', 'sphinx-rtd-theme>=0.5.0', 'nbsphinx>=0.7.0'],
-    'dev': ['coverage>=5.2', 'coverage-badge>=1.0.0', 'pytest-cov>=2.10.0', 'twine>=3.2.0',
-            'ipython>=7.16.0', 'jupyter>=1.0.0', 'notebook>=6.0.0', 'tox>=3.18.0'],
+    'textproc': ['spacy>=3.2.0', 'bidict>=0.21.0', 'loky>=3.0.0'],
+    'textproc_extra': ['PyICU>=2.8', 'nltk>=3.6.0'],
+    'wordclouds': ['wordcloud>=1.8.0,<1.9', 'Pillow>=9.0.0'],
+    'lda': ['lda>=2.0'],
+    'sklearn': ['scikit-learn>=1.0.0'],
+    'gensim': ['gensim>=4.1.0'],
+    'topic_modeling_eval_extra': ['gmpy2>=2.1.0'],
+    'test': ['pytest>=7.0.0', 'hypothesis>=6.36.0'],
+    'doc': ['Sphinx>=4.4.0', 'sphinx-rtd-theme>=1.0.0', 'nbsphinx>=0.8.0'],
+    'dev': ['coverage>=6.3', 'coverage-badge>=1.1.0', 'pytest-cov>=3.0.0', 'twine>=3.8.0',
+            'ipython>=8.0.0', 'jupyter>=1.0.0', 'notebook>=6.4.0', 'tox>=3.24.0', 'setuptools>=60.7.0'],
 }
 
-DEPS_EXTRA['recommended'] = DEPS_EXTRA['excel_export'] + DEPS_EXTRA['wordclouds']
+# DEPS_EXTRA['minimal'] = DEPS_BASE   # doesn't work with extras_require and pip currently
+# see https://github.com/pypa/setuptools/issues/1139
+
+DEPS_EXTRA['recommended'] = DEPS_EXTRA['textproc'] + DEPS_EXTRA['wordclouds']
 DEPS_EXTRA['all'] = []
 for k, deps in DEPS_EXTRA.items():
     if k not in {'recommended', 'all'}:
@@ -71,9 +74,9 @@ setup(
         'Operating System :: OS Independent',
         'Programming Language :: Python',
         'Programming Language :: Python :: 3',
-        'Programming Language :: Python :: 3.6',
-        'Programming Language :: Python :: 3.7',
         'Programming Language :: Python :: 3.8',
+        'Programming Language :: Python :: 3.9',
+        'Programming Language :: Python :: 3.10',
 
         'Topic :: Scientific/Engineering :: Information Analysis',
         'Topic :: Software Development :: Libraries :: Python Modules',
@@ -84,7 +87,7 @@ setup(
 
     packages=find_packages(exclude=['tests', 'examples']),
     include_package_data=True,
-    python_requires='>=3.6',
+    python_requires='>=3.8',
     install_requires=DEPS_BASE,
     extras_require=DEPS_EXTRA
 )
