@@ -6,38 +6,74 @@ CLI module
 Markus Konrad <markus.konrad@wzb.eu>
 """
 
+HELP_TEXT = """tmtoolkit installation setup
+
+Run
+
+    python -m tmtoolkit setup <LANGUAGES>
+
+to install all necessary language models for languages listed as
+comma-separated language ISO codes in <LANGUAGES>. The list must
+be specified without spaces. Example:
+
+    python -m tmtoolkit setup en,de,ru
+
+This will install language models for English, German and Russian.
+To install all available language models, run:
+
+    python -m tmtoolkit setup all
+
+You can pass two additional arguments:
+
+    --variants=...  sets the model size variants to install; default is --variants=sm,md
+    --no-update     if this argument is passed, only models that are not installed so far will be installed
+"""
+
 if __name__ == '__main__':
     import sys
     import subprocess
     import json
 
-    from tmtoolkit.preprocess import DEFAULT_LANGUAGE_MODELS
+    try:
+        from tmtoolkit.corpus import DEFAULT_LANGUAGE_MODELS
+    except ImportError:
+        print('error: tmtoolkit is not installed with the dependencies required for text processing; '
+              'install tmtoolkit with the [recommended] or [textproc] option', file=sys.stderr)
+        exit(1)
 
     def _setup(args):
+        from spacy.cli.download import download
+
+        variants_switch = '--variants='
+        i_variants_arg = None
+        for i, arg in enumerate(args):
+            if arg.startswith(variants_switch):
+                i_variants_arg = i
+                break
+
+        if i_variants_arg is not None:
+            vararg = args.pop(i_variants_arg)
+            variants = vararg[len(variants_switch):].split(',')
+        else:
+            variants = ['sm', 'md']
+
         try:
-            import spacy
-            from spacy.cli.download import download
-        except ImportError:
-            print('error: required package "spacy" is not installed', file=sys.stderr)
-            exit(1)
+            args.remove('--no-update')
+            no_update = True
+        except ValueError:
+            no_update = False
 
         if not args:
             print('error: you must pass a list of two-letter ISO 639-1 language codes to install the respective '
                   'language models or the string "all" to install all available language models', file=sys.stderr)
-            exit(2)
-        else:
-            try:
-                args.pop(args.index('--no-update'))
-                no_update = True
-            except ValueError:
-                no_update = False
+            exit(3)
 
-            if args == ['all']:
-                install_languages = list(DEFAULT_LANGUAGE_MODELS.keys())
-            else:
-                install_languages = []
-                for arg in args:
-                    install_languages.extend([l for l in map(str.strip, arg.split(',')) if l])
+        if args == ['all']:
+            install_languages = list(DEFAULT_LANGUAGE_MODELS.keys())
+        else:
+            install_languages = []
+            for arg in args:
+                install_languages.extend([l for l in map(str.strip, arg.split(',')) if l])
 
         print('checking if required spaCy data packages are installed...')
 
@@ -47,42 +83,50 @@ if __name__ == '__main__':
                                                    '--format', 'json'])
         except subprocess.CalledProcessError as exc:
             print('error: calling pip failed with the following error message\n' + str(exc), file=sys.stderr)
-            exit(3)
+            exit(4)
 
         piplist = json.loads(piplist_str)
         installed_pkgs = set(item['name'] for item in piplist)
-        model_pkgs = dict(zip(DEFAULT_LANGUAGE_MODELS.keys(),
-                              map(lambda x: x.replace('_', '-') + '-sm', DEFAULT_LANGUAGE_MODELS.values())))
 
-        for lang in install_languages:
-            if lang not in DEFAULT_LANGUAGE_MODELS.keys():
-                print('error: no language model for language code "%s"' % lang, file=sys.stderr)
-                exit(4)
+        for modelvar in variants:
+            model_pkgs = dict(zip(DEFAULT_LANGUAGE_MODELS.keys(),
+                                  map(lambda x: x.replace('_', '-') + '-' + modelvar,
+                                      DEFAULT_LANGUAGE_MODELS.values())))
 
-            lang_model_pkg = model_pkgs[lang]
+            for lang in install_languages:
+                if lang not in DEFAULT_LANGUAGE_MODELS.keys():
+                    print(f'error: no language model for language code "{lang}"', file=sys.stderr)
+                    exit(5)
 
-            if no_update and lang_model_pkg in installed_pkgs:
-                print('language model package "%s" for language code "%s" is already installed -- skipping'
-                      % (lang_model_pkg, lang))
-                continue
+                lang_model_pkg = model_pkgs[lang]
 
-            lang_model = DEFAULT_LANGUAGE_MODELS[lang] + '_sm'
-            print('installing language model "%s" for language code "%s"...' % (lang_model, lang))
-            download(lang_model)
+                if no_update and lang_model_pkg in installed_pkgs:
+                    print(f'language model package "{lang_model_pkg}" for language code "{lang}" is already installed '
+                          f'-- skipping')
+                    continue
+
+                lang_model = DEFAULT_LANGUAGE_MODELS[lang] + '_' + modelvar
+                print(f'installing language model "{lang_model}" for language code "{lang}"...')
+                download(lang_model)
 
         print('done.')
 
+    def _help(args):
+        print(HELP_TEXT)
+
     commands = {
-        'setup': _setup
+        'setup': _setup,
+        'help': _help,
     }
 
     if len(sys.argv) <= 1:
         print('available commands: ' + ', '.join(commands.keys()))
-        exit(1)
+        print('run `python -m tmtoolkit help` for help')
+        exit(6)
 
     cmd = sys.argv[1]
     if cmd in commands.keys():
         commands[cmd](sys.argv[2:])
     else:
         print('command not supported:', cmd, file=sys.stderr)
-        exit(2)
+        exit(7)
