@@ -1,11 +1,11 @@
 import numpy as np
+import pandas as pd
 import pytest
 from hypothesis import settings, given, strategies as st
 from scipy.sparse import coo_matrix, csr_matrix, issparse
 
 from ._testtools import strategy_dtm
 
-from tmtoolkit._pd_dt_compat import USE_DT, FRAME_TYPE, pd_dt_colnames
 from tmtoolkit import bow
 
 try:
@@ -14,6 +14,8 @@ try:
 except ImportError:
     GENSIM_INSTALLED = False
 
+pytestmark = [pytest.mark.filterwarnings("ignore:divide by zero"),   # happens due to generated data by hypothesis
+              pytest.mark.filterwarnings("ignore:invalid value")]
 
 @given(
     dtm=strategy_dtm(),
@@ -59,11 +61,17 @@ def test_doc_frequencies(dtm, matrix_type):
         assert df_abs.shape == (dtm_arr.shape[1],)
         assert all([0 <= v <= n_docs for v in df_abs])
 
-        df_rel = bow.bow_stats.doc_frequencies(dtm, proportions=True)
+        df_rel = bow.bow_stats.doc_frequencies(dtm, proportions=1)
         assert isinstance(df_rel, np.ndarray)
         assert df_rel.ndim == 1
         assert df_rel.shape == (dtm_arr.shape[1],)
         assert all([0 <= v <= 1 for v in df_rel])
+
+        df_log = bow.bow_stats.doc_frequencies(dtm, proportions=2)
+        assert isinstance(df_log, np.ndarray)
+        assert df_log.ndim == 1
+        assert df_log.shape == (dtm_arr.shape[1],)
+        assert np.allclose(np.exp(df_log), df_rel)
 
 
 def test_doc_frequencies2():
@@ -81,7 +89,7 @@ def test_doc_frequencies2():
 @given(
     dtm=strategy_dtm(),
     matrix_type=st.integers(min_value=0, max_value=1),
-    proportions=st.booleans()
+    proportions=st.integers(min_value=0, max_value=2)
 )
 def test_codoc_frequencies(dtm, matrix_type, proportions):
     if matrix_type == 1:
@@ -101,21 +109,26 @@ def test_codoc_frequencies(dtm, matrix_type, proportions):
 
     cooc = bow.bow_stats.codoc_frequencies(dtm, proportions=proportions)
 
-    if matrix_type == 1:
+    if matrix_type == 1 and proportions != 2:
         assert issparse(cooc)
+        cooc = cooc.todense()
     else:
         assert isinstance(cooc, np.ndarray)
 
     assert cooc.shape == (n_vocab, n_vocab)
 
-    if matrix_type == 1:
-        cooc = cooc.todense()
-
-    assert np.all(0 <= cooc)
-
-    if proportions:
+    if proportions > 0:
         assert np.all(cooc <= 1)
+
+        if proportions == 1:
+            assert np.all(0 <= cooc)
+        else:   # proportions == 2
+            expected = bow.bow_stats.codoc_frequencies(dtm, proportions=1)
+            if issparse(expected):
+                expected = expected.todense()
+            np.allclose(np.exp(cooc) - 1, expected)
     else:
+        assert np.all(0 <= cooc)
         assert np.all(cooc <= n_docs)
 
 
@@ -153,35 +166,23 @@ def test_term_frequencies(dtm, matrix_type):
         assert tf.shape == (dtm_arr.shape[1],)
         assert tf.tolist() == [sum(row) for row in dtm_arr.T]
 
+        if np.sum(dtm) > 0:
+            tf_prop = bow.bow_stats.term_frequencies(dtm, proportions=1)
+            assert tf_prop.ndim == 1
+            assert tf_prop.shape == (dtm_arr.shape[1],)
+            assert np.all(tf_prop>= 0)
+            assert np.all(tf_prop <= 1)
+            assert np.isclose(tf_prop.sum(), 1.0)
 
-@given(
-    dtm=strategy_dtm(),
-    matrix_type=st.integers(min_value=0, max_value=1)
-)
-def test_term_frequencies_proportions(dtm, matrix_type):
-    if matrix_type == 1:
-        dtm = coo_matrix(dtm)
-        dtm_arr = dtm.A
-        dtm_flat = dtm.A.flatten()
-    else:
-        dtm_arr = dtm
-        dtm_flat = dtm.flatten()
-
-    if dtm.ndim != 2:
-        with pytest.raises(ValueError):
-            bow.bow_stats.term_frequencies(dtm, proportions=True)
-    else:
-        if dtm.sum() == 0:
-            with pytest.raises(ValueError):
-                bow.bow_stats.term_frequencies(dtm, proportions=True)
+            tf_logprop = bow.bow_stats.term_frequencies(dtm, proportions=2)
+            assert tf.ndim == 1
+            assert tf.shape == (dtm_arr.shape[1],)
+            assert np.allclose(np.exp(tf_logprop), tf_prop)
         else:
-            tp = bow.bow_stats.term_frequencies(dtm, proportions=True)
-            assert tp.ndim == 1
-            assert tp.shape == (dtm_arr.shape[1],)
-
-            if len(dtm_flat) > 0:
-                assert np.isclose(tp.sum(), 1.0)
-                assert all(0 <= v <= 1 for v in tp)
+            with pytest.raises(ValueError):
+                bow.bow_stats.term_frequencies(dtm, proportions=1)
+            with pytest.raises(ValueError):
+                bow.bow_stats.term_frequencies(dtm, proportions=2)
 
 
 @given(
@@ -522,7 +523,7 @@ def test_sorted_terms_example():
     top_n=st.integers(min_value=0, max_value=10),
     ascending=st.booleans(),
 )
-def test_sorted_terms_datatable(dtm, matrix_type, lo_thresh, hi_thresh, top_n, ascending):
+def test_sorted_terms_table(dtm, matrix_type, lo_thresh, hi_thresh, top_n, ascending):
     if matrix_type == 1:
         dtm = coo_matrix(dtm)
 
@@ -540,12 +541,13 @@ def test_sorted_terms_datatable(dtm, matrix_type, lo_thresh, hi_thresh, top_n, a
 
     if lo_thresh is not None and hi_thresh is not None and lo_thresh > hi_thresh:
         with pytest.raises(ValueError):
-            bow.bow_stats.sorted_terms_datatable(dtm, vocab, doc_labels, lo_thresh, hi_thresh, top_n, ascending)
+            bow.bow_stats.sorted_terms_table(dtm, vocab, doc_labels, lo_thresh, hi_thresh, top_n, ascending)
     else:
-        res = bow.bow_stats.sorted_terms_datatable(dtm, vocab, doc_labels, lo_thresh, hi_thresh, top_n, ascending)
+        res = bow.bow_stats.sorted_terms_table(dtm, vocab, doc_labels, lo_thresh, hi_thresh, top_n, ascending)
 
-        assert isinstance(res, FRAME_TYPE)
-        assert pd_dt_colnames(res) == ['doc', 'token', 'value']
+        assert isinstance(res, pd.DataFrame)
+        assert res.columns.tolist() == ['token', 'value']
+        assert res.index.names == ['doc', 'rank']
 
 
 @given(
@@ -579,41 +581,6 @@ def test_dtm_to_dataframe(dtm, matrix_type):
     assert np.array_equal(df.to_numpy(), dtm_arr)
     assert np.array_equal(df.index.values, doc_labels)
     assert np.array_equal(df.columns.values, vocab)
-
-
-@given(
-    dtm=strategy_dtm(),
-    matrix_type=st.integers(min_value=0, max_value=1))
-@settings(deadline=1000)
-def test_dtm_to_datatable(dtm, matrix_type):
-    if not USE_DT:
-        pytest.skip('datatable not installed')
-
-    if matrix_type == 1:
-        dtm = coo_matrix(dtm)
-        dtm_arr = dtm.A
-    else:
-        dtm_arr = dtm
-
-    doc_labels = ['doc%d' % i for i in range(dtm.shape[0])]
-    vocab = ['t%d' % i for i in range(dtm.shape[1])]
-
-    # check invalid doc_labels
-    if len(doc_labels) > 0:
-        with pytest.raises(ValueError):
-            bow.dtm.dtm_to_datatable(dtm, doc_labels[:-1], vocab)
-
-    # check invalid vocab
-    if len(vocab) > 0:
-        with pytest.raises(ValueError):
-            bow.dtm.dtm_to_datatable(dtm, doc_labels, vocab[:-1])
-
-    # check with valid doc_labels and vocab
-    df = bow.dtm.dtm_to_datatable(dtm, doc_labels, vocab)
-    assert df.shape == (dtm.shape[0], dtm.shape[1] + 1)  # +1 due to doc column
-    assert np.array_equal(df[:, 0].to_list()[0], doc_labels)
-    assert np.array_equal(pd_dt_colnames(df), ['_doc'] + vocab)
-    assert np.array_equal(df[:, 1:].to_numpy(), dtm_arr)
 
 
 @given(
