@@ -2,7 +2,7 @@
 Bag-of-Words (BoW) sub-package with modules for generating document-term-matrices (DTMs) and some common statistics for
 the BoW model.
 
-Markus Konrad <markus.konrad@wzb.eu>
+.. codeauthor:: Markus Konrad <markus.konrad@wzb.eu>
 """
 
 from . import bow_stats, dtm
