@@ -1,5 +1,7 @@
 """
-Common constants and functions
+Common constants and functions for topic modeling sub-package.
+
+.. codeauthor:: Markus Konrad <markus.konrad@wzb.eu>
 """
 
 
