@@ -3,6 +3,8 @@ Metrics for topic model evaluation.
 
 In order to run model evaluations in parallel use one of the modules :mod:`~tmtoolkit.topicmod.tm_gensim`,
 :mod:`~tmtoolkit.topicmod.tm_lda` or :mod:`~tmtoolkit.topicmod.tm_sklearn`.
+
+.. codeauthor:: Markus Konrad <markus.konrad@wzb.eu>
 """
 
 import numpy as np
@@ -137,9 +139,9 @@ metric_held_out_documents_wallach09.direction = 'maximize'
 
 def metric_cao_juan_2009(topic_word_distrib):
     """
-    Calculate metric as in [Cao2008]_ using topic-word distribution `topic_word_distrib`.
+    Calculate metric as in [Cao2009]_ using topic-word distribution `topic_word_distrib`.
 
-    .. [Cao2008] Cao Juan, Xia Tian, Li Jintao, Zhang Yongdong, and Tang Sheng. 2009. A density-based method for
+    .. [Cao2009] Cao Juan, Xia Tian, Li Jintao, Zhang Yongdong, and Tang Sheng. 2009. A density-based method for
                  adaptive LDA model selection. Neurocomputing — 16th European Symposium on Artificial Neural Networks
                  2008 72, 7–9: 1775–1781. <http://doi.org/10.1016/j.neucom.2008.06.011>.
 
@@ -192,7 +194,8 @@ def metric_arun_2010(topic_word_distrib, doc_topic_distrib, doc_lengths):
     #return entropy(cm1, cm2) + entropy(cm2, cm1)
 
     # use it as in the paper (note: cm1 and cm2 are not prob. distributions that sum up to 1)
-    return np.sum(cm1*np.log(cm1/cm2)) + np.sum(cm2*np.log(cm2/cm1))
+    #return np.sum(cm1*np.log(cm1/cm2)) + np.sum(cm2*np.log(cm2/cm1))
+    return np.sum(cm1 * (np.log(cm1) - np.log(cm2))) + np.sum(cm2 * (np.log(cm2) - np.log(cm1)))
 metric_arun_2010.direction = 'minimize'
 
 
@@ -277,7 +280,7 @@ def metric_coherence_mimno_2011(topic_word_distrib, dtm, top_n=20, eps=1e-12, no
 
         for m in range(1, top_n):
             for l in range(m):
-                c_t += np.log((codf[m, l] + eps) / df[l])
+                c_t += np.log(codf[m, l] + eps) - np.log(df[l])
 
         coh.append(c_t)
 
@@ -428,14 +431,14 @@ def results_by_parameter(res, param, sort_by=None, sort_desc=False):
          ...,
          (parameter_set_n, {'<metric_name>': result_n, ...})])
 
-    Then returns a list with tuple pairs using only the parameter `param` from the parameter sets in the evaluation
-    results such that the returned list is:
+    Then returns a list with tuple pairs using only the *m* parameter(s) listed in `param` from the parameter sets in
+    the evaluation results such that the returned list is:
 
     .. code-block:: text
 
-        [(param_1, {'<metric_name>': result_1, ...}),
+        [(param_1_0, ..., param_1_m, {'<metric_name>': result_1, ...}),
          ...,
-         (param_n, {'<metric_name>': result_n, ...})]
+         (param_n_0, ..., param_n_m, {'<metric_name>': result_n, ...})]
 
     Optionally order either by parameter value (`sort_by` is None - the default) or by result metric
     (``sort_by='<metric name>'``).
@@ -449,21 +452,26 @@ def results_by_parameter(res, param, sort_by=None, sort_desc=False):
     if len(res) == 0:
         return []
 
-    tuples = [(p[param], r) for p, r in res]
+    if isinstance(param, str):
+        param = [param]
 
-    # single validation results
-    if len(tuples[0]) != 2:
-        raise ValueError('invalid evaluation results passed')
+    tuples = []
+    for res_params, res_vals in res:
+        row = tuple(res_params[p] for p in param) + (res_vals,)
+        tuples.append(row)
 
-    params, metric_results = list(zip(*tuples))
-    if sort_by:
-        sorted_ind = argsort([r[sort_by] for r in metric_results])
+    metric_results = list(zip(*tuples))[-1]
+
+    if sort_by is None:
+        sort_by = param[0]
+
+    if sort_by in param:
+        s = param.index(sort_by)
+        sorted_ind = argsort([r[s] for r in tuples])
     else:
-        sorted_ind = argsort(params)
+        sorted_ind = argsort([r[sort_by] for r in metric_results])
 
     if sort_desc:
         sorted_ind = reversed(sorted_ind)
 
-    measurements = tuples
-
-    return [measurements[i] for i in sorted_ind]
+    return [tuples[i] for i in sorted_ind]
